chore(scene.py): move `scenes/scene.py` -> `surfaces/scenes/scene.py`
